\name{gooogleplus-package}
\alias{gooogleplus-package}
\alias{gooogleplus}
\docType{package}
\title{
\packageTitle{gooogleplus}
}
\description{
\packageDescription{gooogleplus}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{gooogleplus}

Maintainer: \packageMaintainer{gooogleplus}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = gooogleplus)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{gooogleplus}}
%\section{Documentation Index}{\packageIndices{gooogleplus}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
